rename 'Situs Luculentus' to 'Robigo Luculenta'
Both 'situs' and 'robigo' (and 'rubigo' as well) can be translated
as 'rust' (the fungus). At first I thought situs is better, but it
is hard to say. I think robigo sounds better, and it keeps the 'r'.
(Rubigo keeps the 'u' as well, but then it sounds like Ruby ...)
Robigo also means rust on metals. Situs does not. Also, robigo is
feminine, so the proper combination is 'robigo luculenta', meaning
'shiny rust' / 'bright rust' / 'impressive rust'.