implement Mul for Vector3
Apparently, only multiplication with `f32` on the right hand side is
allowed. Trying to implement `Mul<Vector3, Vector3> for f32` results
in a long stream of error messages. At this point I do not know why
it is not allowed, but it is not really required, so I will just stick
with scalar multiplication on the right.