rename cargo.toml to Cargo.toml
I prefer to use lowercase names for everything, having a Cargo.toml makes this
an inconsistent mess ... however, Cargo expects a capitalised filename, and
most non-Windows filesystems are case sensitive, so unfortunately it has to
be Cargo.toml. I wonder why Cargo wants an uppercase filename. (It is
customary for make, but make does accept a lowercase makefile as well.)