use uppercase names for statics
This fixes the warnings about uppercase names for statics. Personally, I
dislike all caps for constants. It looks screamy, and it adds little value.
It makes it easy to spot constants, but what is the value of that? It feels
an awful lot like Hungarian notation. Besides, variables are immutable by
default, and there is no distinct naming convention for mutable vs immutable
variables.

Anyway, I dislike all caps, but I do like consistency, so if this is "The
Rust Way", then let's do it that way.