fix and specialise pop_front_iter for libcollections reform
This also removes `pop_iter`, which was not used. The implementation is
not generic over the container any more, but specialised for `RingBuf`.
In this case I can get away with it, because it is only used for
`RingBuf`, but the generic approach was nicer in my opinion. Anyway,
there will probably be more changes in libcollections, so let’s see what
happens.