disable Travis success notifications
Email notifications would only trigger when the status changes from broken to
good, but rust-ci triggers a rebuild every day, so when you do not commit, it
will keep sending a 'build fixed' email every day when the latest commit is one
that fixed a failing build. That is too verbose for me.