fix stack overflow by moving mapped photons to the heap
Apparently the stack size has changed, causing main to crash on a stack overlow.
This moves mapped photons from a fixed-size array on the stack to a vector,
which stores its data on the heap. (I could have used a boxed fixed-size array
because it will never be resized, but the vector suffices.)