upgrade from old_io::sleep to thread::sleep
The new sleep is an unstable feature. An alternative would be to use `yield_now`.
This is stable, and the effect is probably similar, though I expect that
`yield_now` will return much faster than 100ms in general, which could be a bit
wasteful, because most tasks are long-running, so the probability of a new task
becoming available when the call returns is quite low, I think. However, I did
not measure this. A yield might be fine.