remove Travis notification config
I added this back when rust-ci.org was still up. After fixing a build, the
"it has been fixed" message would be re-sent every time the build was triggered
until a new commit was pushed, but now that Rust CI is broken it serves no
purpose anymore. Simpler is better.